feat: add confirmation popup
