+++ packages/core/src/components/pages/FormView/Alerts/index.ts
@@ -16,4 +16,5 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+export {default as ConfirmationPopup} from './ConfirmationPopup';
 export {default as ConstraintsValidatorPopup} from './ConstraintsValidatorPopup';
