+++ packages/core/src/components/pages/FormView/Buttons/index.ts
@@ -1,1 +1,1 @@
-export {default as BasicActionsButton} from './BasicActionsButton';
+export {default as FloatingTools} from './FloatingTools';
